add support to optimization models
