+++ src/components/Navbar/Navbar.jsx
@@ -1,4 +1,4 @@
-// src/components/Navbar.js
+// src/components/Navbar/Navbar.jsx
 import React from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 
@@ -7,7 +7,7 @@ function Navbar() {
 
   const handleLogout = () => {
     localStorage.removeItem('token');
-    navigate('/signin');
+    window.location.reload();
   };
 
   return (
