+++ src/pages/ModelRunner/ModelRunner.css
@@ -1,0 +1,13 @@
+.spinner {
+    animation: spin 1s linear infinite;
+}
+
+@keyframes spin {
+    0% {
+        transform: rotate(0deg);
+    }
+
+    100% {
+        transform: rotate(360deg);
+    }
+}
